Renewed Bet Structure

+++ Node.cs
@@ -9,7 +9,7 @@ namespace LimitTheRisks
         public Node Win2 { get; set; }
         public Node Draw { get; set; }
         public double Coef { get; set; }
-        public double Level { get; set; }
+        public int Level { get; set; }
         public double Prob { get; set; }
         public double LocalCoef { get; set; }
         public double LocalProb { get; set; }
